refactor card, ticket to sync with c5

+++ afc-ops-service/src/main/java/com/vdt/afc_ops_service/integration/level5/dto/response/Level5BusinessSyncResult.java
@@ -18,9 +18,9 @@ public class Level5BusinessSyncResult {
     private String syncType;
     private String correlationId;
     private int processedCount;
-    private int createdCount;
-    private int updatedCount;
-    private int ignoredCount;
-    private int rejectedCount;
+    private Long createdCount;
+    private Long updatedCount;
+    private Long ignoredCount;
+    private Long rejectedCount;
     private List<Level5BusinessSyncItemResult> items;
 }
